fix bugs on testing
