+++ server/package.json
@@ -13,8 +13,7 @@
     "db:seed:dev": "NODE_ENV=development sequelize db:seed:all",
     "db:seed:test": "NODE_ENV=test sequelize db:seed:all",
     "dev": "NODE_ENV=development nodemon bin/http.js",
-    "test": "NODE_ENV=test npx jest",
-    "test-user": "npm run test -- user-login.test.js"
+    "test": "NODE_ENV=test npx jest"
   },
   "keywords": [],
   "author": "",
